requirements.txt
#use pip install -r requirements.txt to install all the dependencies
#used >= for the latest version of the dependency

# Core Framework
fastapi>=0.104.1
uvicorn>=0.34.0
python-multipart>=0.0.6

# LangChain & AI
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.3.7

# Vector Database
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Document Processing
pypdf2>=3.0.1
python-docx>=1.1.0
pytesseract>=0.3.10
pillow>=10.1.0
PyMuPDF>=1.23.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1

# Web Framework
jinja2>=3.1.2
aiofiles>=23.2.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0


#gemini generative ai

google-generativeai>=0.3.0
langchain-google-genai>=0.0.5



sentence-transformers>=2.2.2
chromadb>=0.4.18
langchain-community>=0.0.10
langchain-huggingface>=0.0.5
langchain-chroma>=0.0.5